# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: مكتب مصطفى حمزة - Accounting Office API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoiceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateInvoiceInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: OK
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }
    post:
      operationId: createReport
      tags: [reports]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReportInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
  /reports/{id}:
    delete:
      operationId: deleteReport
      tags: [reports]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: OK
...
components:
  schemas:
...
    InvoiceStatus:
      type: string
      enum: [معلقة, مدفوعة]
    Invoice:
      type: object
      additionalProperties: false
      required: [id, clientName, amount, status, createdAt]
      properties:
        id: { type: integer }
        clientId: { type: integer, nullable: true }
        clientName: { type: string }
        amount: { type: number }
        status: { $ref: "#/components/schemas/InvoiceStatus" }
        description: { type: string, nullable: true }
        addedById: { type: integer, nullable: true }
        addedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateInvoiceInput:
      type: object
      additionalProperties: false
      required: [clientId, amount, status]
      properties:
        clientId: { type: integer }
        amount: { type: number, minimum: 0 }
        status: { $ref: "#/components/schemas/InvoiceStatus" }
        description: { type: string, nullable: true }
    UpdateInvoiceInput:
      type: object
      additionalProperties: false
      properties:
        amount: { type: number, minimum: 0 }
        status: { $ref: "#/components/schemas/InvoiceStatus" }
        description: { type: string, nullable: true }
...
    Report:
      type: object
      additionalProperties: false
      required: [id, title, summary, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        period: { type: string, nullable: true }
        summary: { type: string }
        addedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateReportInput:
      type: object
      additionalProperties: false
      required: [title, summary]
      properties:
        title: { type: string, minLength: 1 }
        period: { type: string, nullable: true }
        summary: { type: string, minLength: 1 }
...